Roles listar corregido.

+++ BackEnd/src/main/java/pe/edu/upc/backend/controllers/RolController.java
@@ -20,6 +20,7 @@ public class RolController {
     public void agregar(@RequestBody RolDTO rolDTO) {
         ModelMapper m = new ModelMapper();
         Rol ciudad = m.map(rolDTO, Rol.class);
+
         rS.insert(ciudad);
     }
     @PutMapping
@@ -36,11 +37,11 @@ public class RolController {
         }).collect(Collectors.toList());
     }
     @DeleteMapping("/{id}")
-    public void eliminar(@PathVariable("id") int id) {
+    public void eliminar(@PathVariable("id") Long id) {
         rS.delete(id);
     }
     @GetMapping("/{id}")
-    public RolDTO listById(@PathVariable("id") int id) {
+    public RolDTO listById(@PathVariable("id") Long id) {
         ModelMapper m = new ModelMapper();
         return m.map(rS.listById(id),RolDTO.class);
     }
