+++ BackEnd/src/main/java/pe/edu/upc/backend/repositories/IRolRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.backend.entities.Rol;
 import java.util.List;
 @Repository
-public interface IRolRepository extends JpaRepository<Rol, Integer> {
+public interface IRolRepository extends JpaRepository<Rol, Long> {
     @Query(value = "select c.rol, count(*)\n" +
             " from rol c inner join usuario us\n" +
             " on c.user_id= us.id_usuario\n" +
